refactor: organize models and types
