+++ Typeform.Tests/AnswerDeserialization.cs
@@ -9,7 +9,7 @@ public class AnswerDeserializationTests
   private T GetAnswerFromFixture<T>(int answerIndex, int itemIndex = 0) where T : TypeformAnswer
   {
     var responsesFixture = FixturesHelper.GetResponsesFixture();
-    var responses = JsonSerializer.Deserialize<TypeformResponsesContainer>(responsesFixture, TypeformClient.DefaultSystemTextJsonSerializerOptions);
+    var responses = JsonSerializer.Deserialize<TypeformResponseItems>(responsesFixture, TypeformClient.DefaultSystemTextJsonSerializerOptions);
 
     return responses!.Items[itemIndex].Answers.GetAnswer<T>(answerIndex);
   }
@@ -19,7 +19,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformTextAnswer>(0);
 
-    Assert.Equal(AnswerType.Text, answer.Type);
+    Assert.Equal(TypeformAnswerType.Text, answer.Type);
     Assert.Equal("Job opportunities", answer.Text);
   }
 
@@ -28,7 +28,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformBooleanAnswer>(1);
 
-    Assert.Equal(AnswerType.Boolean, answer.Type);
+    Assert.Equal(TypeformAnswerType.Boolean, answer.Type);
     Assert.False(answer.Boolean);
   }
 
@@ -37,7 +37,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformEmailAnswer>(4);
 
-    Assert.Equal(AnswerType.Email, answer.Type);
+    Assert.Equal(TypeformAnswerType.Email, answer.Type);
     Assert.Equal("lian1078@other.com", answer.Email);
   }
 
@@ -46,7 +46,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformNumberAnswer>(5);
 
-    Assert.Equal(AnswerType.Number, answer.Type);
+    Assert.Equal(TypeformAnswerType.Number, answer.Type);
     Assert.Equal(1, answer.Number);
   }
 
@@ -55,7 +55,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformChoicesAnswer>(10);
 
-    Assert.Equal(AnswerType.Choices, answer.Type);
+    Assert.Equal(TypeformAnswerType.Choices, answer.Type);
     Assert.NotNull(answer.Choices);
     Assert.NotEmpty(answer.Choices.Labels);
     Assert.Contains("New York", answer.Choices.Labels);
@@ -67,7 +67,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformDateAnswer>(11);
 
-    Assert.Equal(AnswerType.Date, answer.Type);
+    Assert.Equal(TypeformAnswerType.Date, answer.Type);
     var expectedDate = new DateTime(2012, 3, 20, 0, 0, 0, DateTimeKind.Utc);
     Assert.Equal(expectedDate, answer.Date);
   }
@@ -77,7 +77,7 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformChoiceAnswer>(12);
 
-    Assert.Equal(AnswerType.Choice, answer.Type);
+    Assert.Equal(TypeformAnswerType.Choice, answer.Type);
     Assert.NotNull(answer.Choice);
     Assert.Equal("A friend's experience in Sydney", answer.Choice.Label);
   }
@@ -88,7 +88,7 @@ public class AnswerDeserializationTests
     var answer = GetAnswerFromFixture<TypeformFileUrlAnswer>(1, itemIndex: 1);
     var expectedUri = new Uri("https://api.typeform.com/forms/lT9Z2j/responses/7f46165474d11ee5836777d85df2cdab/fields/X4BgU2f1K6tG/files/afd8258fd453-aerial_view_rural_city_latvia_valmiera_urban_district_48132860.jpg");
 
-    Assert.Equal(AnswerType.FileUrl, answer.Type);
+    Assert.Equal(TypeformAnswerType.FileUrl, answer.Type);
     Assert.Equal(expectedUri, answer.FileUrl);
   }
 
@@ -98,7 +98,7 @@ public class AnswerDeserializationTests
     var answer = GetAnswerFromFixture<TypeformUrlAnswer>(15);
     var expectedUri = new Uri("https://www.google.com");
 
-    Assert.Equal(AnswerType.Url, answer.Type);
+    Assert.Equal(TypeformAnswerType.Url, answer.Type);
     Assert.Equal(expectedUri, answer.Url);
   }
 
@@ -113,7 +113,7 @@ public class AnswerDeserializationTests
       Last4 = "1234"
     };
 
-    Assert.Equal(AnswerType.Payment, answer.Type);
+    Assert.Equal(TypeformAnswerType.Payment, answer.Type);
     Assert.Equal(expectedPaymentData.Amount, answer.Payment.Amount);
     Assert.Equal(expectedPaymentData.Name, answer.Payment.Name);
     Assert.Equal(expectedPaymentData.Last4, answer.Payment.Last4);
@@ -124,6 +124,6 @@ public class AnswerDeserializationTests
   {
     var answer = GetAnswerFromFixture<TypeformAnswer>(2);
 
-    Assert.Equal(AnswerType.Unknown, answer.Type);
+    Assert.Equal(TypeformAnswerType.Unknown, answer.Type);
   }
 }
