+++ Typeform.Tests/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using Xunit;
+global using System.Threading.Tasks;
+global using Typeform.Models;
