--- Typeform.Tests/Globals.cs
@@ -1,2 +1,0 @@
-global using Xunit;
-global using System.Threading.Tasks;
