+++ Typeform/GlobalUsings.cs
@@ -1,3 +1,4 @@
 global using System.Text;
 global using System.Text.Json;
 global using Refit;
+global using Typeform.Models;
