+++ Typeform/Models/AnswerType.cs
@@ -1,0 +1,19 @@
+using System.Runtime.Serialization;
+
+namespace Typeform;
+
+public enum TypeformAnswerType
+{
+  Unknown,
+  Text,
+  Choice,
+  Choices,
+  Email,
+  Url,
+  [EnumMember(Value = "file_url")]
+  FileUrl,
+  Boolean,
+  Number,
+  Date,
+  Payment
+}
