+++ Typeform/Models/ResponseAnswers.cs
@@ -1,0 +1,9 @@
+namespace Typeform.Models;
+
+public class TypeformResponseAnswers : List<TypeformAnswer>
+{
+  public T GetAnswer<T>(int index) where T : TypeformAnswer
+  {
+    return (T)this[index];
+  }
+}
