+++ Typeform/Models/ResponseCalculatedFields.cs
@@ -1,0 +1,6 @@
+namespace Typeform.Models;
+
+public class TypeformResponseCalculatedFields
+{
+  public int? Score { get; set; }
+}
