+++ Typeform/Models/Variable.cs
@@ -1,0 +1,8 @@
+namespace Typeform.Models;
+
+public class TypeformVariable
+{
+  public string Key { get; set; }
+
+  public TypeformVariableType Type { get; set; }
+}
