+++ Typeform/Models/VariableNumber.cs
@@ -1,0 +1,6 @@
+namespace Typeform.Models;
+
+public class TypeformVariableNumber : TypeformVariable
+{
+  public int? Number { get; set; }
+}
