+++ Typeform/Models/VariableText.cs
@@ -1,0 +1,6 @@
+namespace Typeform.Models;
+
+public class TypeformVariableText : TypeformVariable
+{
+  public string Text { get; set; }
+}
