+++ Typeform/Models/VariableType.cs
@@ -1,0 +1,10 @@
+namespace Typeform.Models;
+
+public enum TypeformVariableType
+{
+  Unknown,
+
+  Number,
+
+  Text,
+}
