+++ Typeform/TypeformApi.cs
@@ -1,5 +1,3 @@
-using System.Runtime.Serialization;
-
 namespace Typeform;
 
 public interface ITypeformApi
@@ -10,7 +8,7 @@ public interface ITypeformApi
   /// <param name="formId">Unique ID for the form. Find in your form URL. For example, in the URL "https://mysite.typeform.com/to/u6nXL7" the form_id is u6nXL7.</param>
   /// <returns></returns>
   [Get("/forms/{form_id}/responses")]
-  Task<TypeformResponsesContainer> GetFormResponsesAsync(
+  Task<TypeformResponseItems> GetFormResponsesAsync(
     [Authorize("Bearer")] string accessToken,
     [AliasAs("form_id")] string formId);
 
@@ -21,373 +19,10 @@ public interface ITypeformApi
   /// <param name="queryParams">Optional query parameters to pass to Responses endpoint</param>
   /// <returns></returns>
   [Get("/forms/{form_id}/responses")]
-  Task<TypeformResponsesContainer> GetFormResponsesAsync(
+  Task<TypeformResponseItems> GetFormResponsesAsync(
     [Authorize("Bearer")] string accessToken,
     [AliasAs("form_id")] string formId,
     TypeformResponsesParameters queryParams);
-}
-
-public class TypeformResponsesContainer
-{
-
-  public TypeformResponsesContainer()
-  {
-    Items = new List<TypeformResponse>();
-  }
-
-  /// <summary>
-  /// Total number of items in the retrieved collection.
-  /// </summary>
-  /// <value></value>
-  public int TotalItems { get; set; }
-
-  /// <summary>
-  /// Number of pages.
-  /// </summary>
-  /// <value></value>
-  public int PageCount { get; set; }
-
-  /// <summary>
-  /// Array of Typeform Response objects
-  /// </summary>
-  /// <value></value>
-  public IList<TypeformResponse> Items { get; set; }
-}
-
-public class TypeformResponse
-{
-  public TypeformResponse()
-  {
-    Answers = new AnswerList();
-    Variables = new TypeformVariablesList();
-    Hidden = new TypeformDocumentData();
-    Metadata = new TypeformResponseMetadata();
-    Calculated = new TypeformResponseCalculatedFields();
-  }
-
-  /// <summary>
-  /// Unique ID for the response. Note that `response_id` values are unique per form but are not unique globally.
-  /// </summary>
-  /// <value></value>
-  public string ResponseId { get; set; }
-
-  public string LandingId { get; set; }
-
-  /// <summary>
-  /// Time of the form landing. In ISO 8601 format, UTC time, to the second, with T as a delimiter between the date and time
-  /// </summary>
-  /// <value></value>
-  public DateTime LandedAt { get; set; }
-
-  /// <summary>
-  /// Time that the form response was submitted. In ISO 8601 format, UTC time, to the second, with T as a delimiter between the date and time.
-  /// </summary>
-  /// <value></value>
-  public DateTime SubmittedAt { get; set; }
-
-  public string Token { get; set; }
-
-  public TypeformResponseCalculatedFields Calculated { get; set; }
-
-  public TypeformDocumentData Hidden { get; set; }
-
-  public AnswerList Answers { get; set; }
-
-  public TypeformVariablesList Variables { get; set; }
-
-  /// <summary>
-  /// Metadata about a client's HTTP request.
-  /// </summary>
-  /// <value></value>
-  public TypeformResponseMetadata Metadata { get; set; }
-}
-
-public class TypeformResponseCalculatedFields
-{
-  public int? Score { get; set; }
-}
-
-public class TypeformResponseMetadata
-{
-  public string Browser { get; set; }
-
-  /// <summary>
-  /// IP of the client
-  /// </summary>
-  /// <value></value>
-  public string NetworkId { get; set; }
-
-  /// <summary>
-  /// Derived from user agent
-  /// </summary>
-  /// <value></value>
-  public string Platform { get; set; }
-
-  public string Referer { get; set; }
-
-  public string UserAgent { get; set; }
-}
-
-public class TypeformDocumentData : Dictionary<string, JsonElement>
-{
-  public T Value<T>(string key)
-  {
-    if (typeof(T).IsAssignableFrom(typeof(String)))
-    {
-      return (T)(object)this[key].GetString();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Boolean)))
-    {
-      return (T)(object)this[key].GetBoolean();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Int32)))
-    {
-      return (T)(object)this[key].GetInt32();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Int16)))
-    {
-      return (T)(object)this[key].GetInt16();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Int64)))
-    {
-      return (T)(object)this[key].GetInt64();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(DateTime)))
-    {
-      return (T)(object)this[key].GetDateTime();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(DateTimeOffset)))
-    {
-      return (T)(object)this[key].GetDateTimeOffset();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Guid)))
-    {
-      return (T)(object)this[key].GetGuid();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(SByte)))
-    {
-      return (T)(object)this[key].GetSByte();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Single)))
-    {
-      return (T)(object)this[key].GetSingle();
-    }
-    else if (typeof(T).IsAssignableFrom(typeof(Decimal)))
-    {
-      return (T)(object)this[key].GetDecimal();
-    }
-
-    throw new InvalidOperationException($"Unsupported type {typeof(T)} to deserialize from JSON");
-  }
-}
-
-public class AnswerList : List<TypeformAnswer>
-{
-  public T GetAnswer<T>(int index) where T : TypeformAnswer
-  {
-    return (T)this[index];
-  }
-}
-
-public class TypeformVariablesList : List<TypeformVariable>
-{
-  public T GetVariable<T>(int index) where T : TypeformVariable
-  {
-    return (T)this[index];
-  }
-
-  public T GetVariable<T>(string key) where T : TypeformVariable
-  {
-    return (T)this.FirstOrDefault(v => v.Key == key);
-  }
-}
-
-public enum AnswerType
-{
-  Unknown,
-  Text,
-  Choice,
-  Choices,
-  Email,
-  Url,
-  [EnumMember(Value = "file_url")]
-  FileUrl,
-  Boolean,
-  Number,
-  Date,
-  Payment
-}
-
-public enum QuestionType
-{
-  Matrix,
-  Ranking,
-  Date,
-  Dropdown,
-  Email,
-  [EnumMember(Value = "file_upload")]
-  FileUpload,
-  Group,
-  Legal,
-  [EnumMember(Value = "long_text")]
-  LongText,
-  [EnumMember(Value = "multiple_choice")]
-  MultipleChoice,
-  Number,
-  [EnumMember(Value = "opinion_scale")]
-  OpinionScale,
-  Payment,
-  [EnumMember(Value = "picture_choice")]
-  PictureChoice,
-  Rating,
-  [EnumMember(Value = "short_text")]
-  ShortText,
-  Statement,
-  Website,
-  [EnumMember(Value = "yes_no")]
-  YesNo,
-  [EnumMember(Value = "phone_number")]
-  PhoneNumber
-}
-
-public enum TypeformVariableType
-{
-  Unknown,
-
-  Number,
-
-  Text,
-}
-
-public class TypeformVariable
-{
-  public string Key { get; set; }
-
-  public TypeformVariableType Type { get; set; }
-}
-
-public class TypeformVariableText : TypeformVariable
-{
-  public string Text { get; set; }
-}
-
-public class TypeformVariableNumber : TypeformVariable
-{
-  public int? Number { get; set; }
-}
-
-public class TypeformAnswer
-{
-  public TypeformAnswerField Field { get; set; }
-
-  public AnswerType Type { get; set; }
-}
-
-public class TypeformTextAnswer : TypeformAnswer
-{
-  public string Text { get; set; }
-}
-
-public class TypeformBooleanAnswer : TypeformAnswer
-{
-  public bool Boolean { get; set; }
-}
-
-public class TypeformEmailAnswer : TypeformAnswer
-{
-  public string Email { get; set; }
-}
-
-public class TypeformFileUrlAnswer : TypeformAnswer
-{
-  public Uri FileUrl { get; set; }
-}
-
-public class TypeformUrlAnswer : TypeformAnswer
-{
-  public Uri Url { get; set; }
-}
-
-public class TypeformNumberAnswer : TypeformAnswer
-{
-  public int? Number { get; set; }
-}
-
-public class TypeformChoicesAnswer : TypeformAnswer
-{
-  public TypeformChoicesLabels Choices { get; set; }
-}
-
-public class TypeformChoicesLabels
-{
-  public TypeformChoicesLabels()
-  {
-    Labels = new List<string>();
-  }
-  public IList<string> Labels { get; set; }
-
-  public string Other { get; set; }
-}
-
-public class TypeformChoiceAnswer : TypeformAnswer
-{
-  public TypeformChoiceLabel Choice { get; set; }
-}
-
-public class TypeformChoiceLabel
-{
-  public string Label { get; set; }
-
-  public string Other { get; set; }
-}
-
-public class TypeformDateAnswer : TypeformAnswer
-{
-  public DateTime? Date { get; set; }
-}
-
-public class TypeformPaymentAnswer : TypeformAnswer
-{
-
-  public TypeformPaymentAnswerData Payment { get; set; }
-}
-
-public class TypeformPaymentAnswerData
-{
-  public string Amount { get; set; }
-
-  public string Last4 { get; set; }
-
-  public string Name { get; set; }
-}
-
-public class TypeformAnswerField
-{
-  /// <summary>
-  /// The unique id of the form field the answer refers to.
-  /// </summary>
-  /// <value></value>
-  public string Id { get; set; }
-
-  /// <summary>
-  /// The reference for the question the answer relates to. Use the `ref` value to 
-  /// match answers with questions. The Responses payload only includes `ref` 
-  /// for the fields where you specified them when you created the form.
-  /// </summary>
-  /// <value></value>
-  public string Ref { get; set; }
-
-  /// <summary>
-  /// The field's type in the original form.
-  /// </summary>
-  /// <value></value>
-  public string Type { get; set; }
-
-  /// <summary>
-  /// The form field's title which the answer is related to.
-  /// </summary>
-  /// <value></value>
-  public string Title { get; set; }
 }
 
 public class TypeformResponsesParameters
